feat(main/placelistfactory): move file to your folder
