+++ tests/main/factories/presentation/placeListFactory.spec.tsx
@@ -1,26 +1,17 @@
 import React from 'react';
-import { Place } from 'src/presentation/components/cardList';
 import { render } from '@testing-library/react-native';
+import { Routes } from '~/main/navigation';
 import PlaceList from '../../../../src/presentation/placeList';
+import PlaceListFactory from '../../../../src/main/factories/presentation/placeListFactory';
 
 describe('Main: PlaceListFactory', () => {
   test('should factory the PlaceList with success', () => {
-    const { UNSAFE_getByType } = render(<PlaceListFactory />);
+    const { UNSAFE_getByType } = render(
+      <PlaceListFactory
+        route={{ name: Routes.HOME, key: '', params: undefined }}
+        navigation={undefined}
+      />,
+    );
     expect(UNSAFE_getByType(PlaceList));
   });
 });
-
-const PlaceListFactory = () => {
-  const viewModel = usePlaceList();
-  return <PlaceList {...viewModel} />;
-};
-
-type PlaceListViewModel = {
-  list: Array<Place>;
-  favorite: () => void;
-  showMoreDetails: (place: Place) => void;
-};
-
-const usePlaceList = (): PlaceListViewModel => {
-  return { favorite: () => {}, list: [], showMoreDetails: () => {} };
-};
